\name{Downsample}
\alias{Downsample}
\title{Marker-free annotation: Downsampling}
\description{Downsamples query and database AverageExpression dataframes within a SAHA object to contain only those genes which are shared between the datasets.}
\usage{
Downsample(ann)
}
\arguments{
  \item{ann}{A SAHA object which has been passed through Initialize_MarkerFree() for marker-free annotation.}
}