\name{Initialize_MarkerFree}
\alias{Initialize_MarkerFree}
\title{Marker-free annotation: Initialization}
\description{
Used to clean query and database inputs prior to marker-free annotation. 
}
\usage{
Initialize_MarkerFree(ann)
}
\arguments{
  \item{ann}{A SAHA object created using Create_SAHA_object(), with data_type = "AvgExp" and an AverageExpression dataframe as input.}
}